HiddenBackdoorNMT-master/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
apex=0.1=pypi_0
argon2-cffi=20.1.0=py38h7b6447c_1
astunparse=1.6.3=pypi_0
async_generator=1.10=py_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
bert-score=0.3.6=pypi_0
blas=1.0=mkl
bleach=3.2.1=py_0
bpemb=0.3.2=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
ca-certificates=2020.12.8=h06a4308_0
cachetools=4.2.0=pypi_0
certifi=2020.12.5=py38h06a4308_0
cffi=1.14.4=py38h261ae71_0
chardet=4.0.0=py38h06a4308_1003
click=7.1.2=pypi_0
cloudpickle=1.6.0=pypi_0
cryptography=3.3.1=py38h3c74f83_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py38_0
cython=0.29.21=pypi_0
dataclasses=0.6=pypi_0
datasets=1.1.2=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
deprecated=1.2.10=pypi_0
dill=0.3.3=pypi_0
docopt=0.6.2=pypi_0
editdistance=0.5.3=pypi_0
entrypoints=0.3=py38_0
expat=2.2.10=he6710b0_2
fairseq=0.10.0=pypi_0
fastbpe=0.1.0=pypi_0
filelock=3.0.12=pypi_0
flair=0.6.1.post1=pypi_0
flatbuffers=1.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
ftfy=5.8=pypi_0
future=0.18.2=pypi_0
gast=0.3.3=pypi_0
gdown=3.12.2=pypi_0
gensim=3.8.3=pypi_0
glib=2.66.1=h92f7085_0
google-auth=1.24.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.32.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=pypi_0
hydra-core=1.0.4=pypi_0
hyperopt=0.2.5=pypi_0
icu=58.2=he6710b0_3
idna=2.10=py_0
importlib-metadata=2.0.0=py_1
importlib-resources=3.3.0=pypi_0
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.19.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_0
janome=0.4.1=pypi_0
jedi=0.18.0=py38h06a4308_0
jieba=0.42.1=pypi_0
jinja2=2.11.2=py_0
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py38h06a4308_0
jupyterlab=2.2.6=py_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=1.2.0=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras=2.4.3=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.0=py38h2531618_0
konoha=4.6.2=pypi_0
langdetect=1.0.8=pypi_0
language-tool-python=2.4.6=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
lemminflect=0.2.1=pypi_0
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lru-dict=1.1.6=pypi_0
lxml=4.6.1=pypi_0
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=pypi_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.3.2=h06a4308_0
matplotlib-base=3.3.2=py38h817c723_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
more-itertools=8.6.0=pypi_0
mpld3=0.3=pypi_0
multiprocess=0.70.11.1=pypi_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
networkx=2.5=pypi_0
ninja=1.10.2=py38hff7bd54_0
nltk=3.5=pypi_0
notebook=6.1.6=py38h06a4308_0
num2words=0.5.10=pypi_0
numpy=1.18.5=pypi_0
numpy-base=1.19.2=py38hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
omegaconf=2.0.5=pypi_0
openssl=1.1.1i=h27cfd23_0
opt-einsum=3.3.0=pypi_0
overrides=3.0.0=pypi_0
packaging=20.8=pyhd3eb1b0_0
pandas=1.2.0=py38ha9443f7_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.7.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=py38he98fc37_0
pip=20.3.3=py38h06a4308_0
portalocker=2.0.0=pypi_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyarrow=2.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pygments=2.7.3=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
pytorch=1.7.0=py3.8_cuda11.0.221_cudnn8.0.3_0
pytz=2020.5=pyhd3eb1b0_0
pyyaml=5.3.1=pypi_0
pyzmq=20.0.0=py38h2531618_1
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
regex=2020.10.28=pypi_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
sacrebleu=1.4.14=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.4.1=pypi_0
segtok=1.5.10=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
sentencepiece=0.1.94=pypi_0
setuptools=51.0.0=py38h06a4308_2
sip=4.19.13=py38he6710b0_0
six=1.15.0=py38h06a4308_0
smart-open=3.0.0=pypi_0
sqlite=3.33.0=h62c20be_0
sqlitedict=1.7.0=pypi_0
subword-nmt=0.3.7=pypi_0
tabulate=0.8.7=pypi_0
tensorboard=2.4.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorboardx=2.1=pypi_0
tensorflow=2.4.0=pypi_0
tensorflow-estimator=2.4.0=pypi_0
termcolor=1.1.0=py38_1
terminado=0.9.2=py38h06a4308_0
terminaltables=3.1.0=pypi_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tokenizers=0.9.2=pypi_0
torchaudio=0.7.0=py38
torchtext=0.8.0=pypi_0
torchvision=0.8.1=py38_cu110
tornado=6.1=py38h27cfd23_0
tqdm=4.49.0=pypi_0
traitlets=5.0.5=py_0
transformers=3.4.0=pypi_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
word2number=1.1=pypi_0
wrapt=1.12.1=pypi_0
xxhash=2.0.0=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
